requirements.txt
langchain
python-dotenv
langchain_community
faiss-cpu
streamlit
langchain_openai
openpyxl